feat(linter): configure linter and fix reported problems
