+++ package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "test": "jest 'src/test/'",
     "coverage": "jest 'src/test/' --coverage",
-    "build": "tsc"
+    "build": "tsc",
+    "lint": "eslint ./src/**/*.{ts,js}"
   },
   "keywords": [
     "cqrs",
@@ -23,10 +24,14 @@
     "@types/faker": "^4.1.9",
     "@types/jest": "^25.1.2",
     "@types/node": "^13.7.4",
+    "@typescript-eslint/eslint-plugin": "^2.21.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-airbnb-typescript": "^7.0.0",
     "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^23.7.0",
+    "eslint-plugin-jest": "^23.6.0",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.18.3",
     "faker": "^4.1.0",
     "jest": "^25.1.0",
     "jest-each": "^25.1.0",
