+++ src/index.ts
@@ -1,5 +1,21 @@
-export * from './lib/command-bus';
-export * from './lib/query-bus';
-export * from './lib/message-bus';
-export * from './lib/registry';
-export * from './lib/error';
+import CommandBus from './lib/command/CommandBus';
+import CommandHandler from './lib/command/CommandHandler';
+import QueryBus from './lib/query/QueryBus';
+import QueryHandler from './lib/query/QueryHandler';
+import Middleware from './lib/message/Middleware';
+import HandlerRegistry from './lib/registry/HandlerRegistry';
+import InMemoryHandlerRegistry from './lib/registry/InMemoryHandlerRegistry';
+import HandlerAlreadyExists from './lib/error/HandlerAlreadyExists';
+import HandlerNotFound from './lib/error/HandlerNotFound';
+
+export {
+  CommandBus,
+  CommandHandler,
+  QueryBus,
+  QueryHandler,
+  Middleware,
+  HandlerRegistry,
+  InMemoryHandlerRegistry,
+  HandlerAlreadyExists,
+  HandlerNotFound,
+};
