+++ src/lib/command/CommandHandler.ts
@@ -1,0 +1,3 @@
+export default interface CommandHandler<TCommand, TResult> {
+  execute(cmd: TCommand): TResult;
+}
