--- src/lib/error.ts
@@ -1,2 +1,0 @@
-export class HandlerNotFound extends Error {}
-export class HandlerAlreadyExists extends Error {}
