+++ src/lib/error/HandlerAlreadyExists.ts
@@ -1,0 +1,1 @@
+export default class HandlerAlreadyExists extends Error {}
