+++ src/lib/error/HandlerNotFound.ts
@@ -1,0 +1,1 @@
+export default class HandlerNotFound extends Error {}
