+++ src/lib/message/HandlerFunc.ts
@@ -1,0 +1,3 @@
+export default interface HandlerFunc {
+  <TMsg, TResult>(message: TMsg): TResult | null;
+}
