+++ src/lib/message/Middleware.ts
@@ -1,0 +1,5 @@
+import HandlerFunc from './HandlerFunc';
+
+export default interface Middleware {
+  (handler: HandlerFunc): HandlerFunc;
+}
