+++ src/lib/query/QueryHandler.ts
@@ -1,0 +1,3 @@
+export default interface QueryHandler<TQuery, TResult> {
+  query(query: TQuery): TResult;
+}
