+++ src/lib/registry/HandlerRegistry.ts
@@ -1,0 +1,8 @@
+export interface Class {
+  name: string;
+}
+
+export default interface HandlerRegistry<HandlerType> {
+  locate<T>(cmd: T): HandlerType;
+  register(cmd: Class, handler: HandlerType): HandlerRegistry<HandlerType>;
+}
