--- src/test/mock/index.ts
@@ -1,5 +1,0 @@
-export const delay = <T>(timer: number, callback: Function): Promise<T> => {
-    return new Promise(resolve => setTimeout(() => {
-        resolve(callback());
-    }, timer));
-};
