+++ src/test/mock/message/Message.ts
@@ -1,0 +1,5 @@
+export default class Message {
+  public constructor(
+    public count: number = 0
+  ) {}
+}
