+++ src/test/mock/message/SubMessageBus.ts
@@ -1,0 +1,8 @@
+import MessageBus from '../../../lib/message/MessageBus';
+import HandlerFunc from '../../../lib/message/HandlerFunc';
+
+export default class StubMessageBus extends MessageBus {
+  public execute<TMsg>(message: TMsg, handler: HandlerFunc): void {
+    return this.handle(message, handler);
+  }
+}
