+++ src/test/mock/query/Multiplier.ts
@@ -1,0 +1,6 @@
+export default class Multiplier {
+  public constructor(
+    public readonly value: number,
+    public readonly factor: number,
+  ) {}
+}
