+++ src/test/unit/registry.test.ts
@@ -1,5 +1,8 @@
-import { HandlerRegistry, InMemoryHandlerRegistry } from '../..';
-import { HandlerAlreadyExists, HandlerNotFound } from '../..';
+// eslint-disable-next-line max-classes-per-file
+import {
+  HandlerRegistry, InMemoryHandlerRegistry, HandlerAlreadyExists, HandlerNotFound
+} from '../..';
+
 
 class GenericCommand {}
 class GenericCommand2 {}
@@ -7,44 +10,44 @@ class GenericHandler {}
 class GenericHandler2 {}
 
 describe('InMemoryHandlerRegistry', () => {
-    let registry: HandlerRegistry<GenericHandler>;
+  let registry: HandlerRegistry<GenericHandler>;
 
-    beforeEach(() => {
-        registry = new InMemoryHandlerRegistry<GenericHandler>();
-    });
+  beforeEach(() => {
+    registry = new InMemoryHandlerRegistry<GenericHandler>();
+  });
 
-    it('Should register a new handler', () => {
-        const handler = new GenericHandler();
-        const cmd = new GenericCommand();
+  it('Should register a new handler', () => {
+    const handler = new GenericHandler();
+    const cmd = new GenericCommand();
 
-        registry.register(GenericCommand, handler);
-        expect(registry.locate(cmd)).toBe(handler);
-    });
+    registry.register(GenericCommand, handler);
+    expect(registry.locate(cmd)).toBe(handler);
+  });
 
-    it('Should fail when trying to register a handler for an existing command', () => {
-        const handler = new GenericHandler();
-        const cmd = new GenericCommand();
+  it('Should fail when trying to register a handler for an existing command', () => {
+    const handler = new GenericHandler();
+    const cmd = new GenericCommand();
 
-        registry.register(GenericCommand, handler);
-        expect(() => registry.register(GenericCommand, handler)).toThrowError(HandlerAlreadyExists);
-        expect(registry.locate(cmd)).toBe(handler);
-    });
+    registry.register(GenericCommand, handler);
+    expect(() => registry.register(GenericCommand, handler)).toThrowError(HandlerAlreadyExists);
+    expect(registry.locate(cmd)).toBe(handler);
+  });
 
-    it('Should fail when trying to find a non registered handler', () => {
-        const cmd = new GenericCommand();
-        expect(() => registry.locate(cmd)).toThrowError(HandlerNotFound);
-    });
+  it('Should fail when trying to find a non registered handler', () => {
+    const cmd = new GenericCommand();
+    expect(() => registry.locate(cmd)).toThrowError(HandlerNotFound);
+  });
 
-    it('Should allow multiple commands and handlers', () => {
-        const cmd1 = new GenericCommand();
-        const cmd2 = new GenericCommand2();
-        const handler1 = new GenericHandler();
-        const handler2 = new GenericHandler2();
+  it('Should allow multiple commands and handlers', () => {
+    const cmd1 = new GenericCommand();
+    const cmd2 = new GenericCommand2();
+    const handler1 = new GenericHandler();
+    const handler2 = new GenericHandler2();
 
-        registry.register(GenericCommand, handler1);
-        registry.register(GenericCommand2, handler2);
+    registry.register(GenericCommand, handler1);
+    registry.register(GenericCommand2, handler2);
 
-        expect(registry.locate(cmd1)).toBe(handler1);
-        expect(registry.locate(cmd2)).toBe(handler2);
-    });
+    expect(registry.locate(cmd1)).toBe(handler1);
+    expect(registry.locate(cmd2)).toBe(handler2);
+  });
 });
