--- tslint.json
@@ -1,56 +1,0 @@
-{
-  "extends": "tslint:recommended",
-  "rules": {
-    "array-type": [
-      true,
-      "array"
-    ],
-    "ban-types": false,
-    "comment-format": false,
-    "completed-docs": false,
-    "cyclomatic-complexity": false,
-    "interface-name": false,
-    "linebreak-style": false,
-    "max-classes-per-file": false,
-    "max-file-line-count": false,
-    "max-line-length": [
-      true,
-      140
-    ],
-    "member-ordering": false,
-    "newline-before-return": false,
-    "newline-per-chained-call": false,
-    "no-any": false,
-    "no-empty-interface": false,
-    "no-floating-promises": false,
-    "no-implicit-dependencies": [
-      true,
-      "dev"
-    ],
-    "no-import-side-effect": false,
-    "no-inferred-empty-object-type": false,
-    "no-magic-numbers": false,
-    "no-namespace": false,
-    "no-null-keyword": false,
-    "no-parameter-properties": false,
-    "no-submodule-imports": false,
-    "no-unbound-method": false,
-    "no-unnecessary-class": false,
-    "no-unnecessary-qualifier": false,
-    "no-unsafe-any": false,
-    "no-reference": false,
-    "no-void-expression": false,
-    "only-arrow-functions": false,
-    "prefer-function-over-method": false,
-    "prefer-template": false,
-    "prefer-readonly": false,
-    "promise-function-async": false,
-    "space-before-function-paren": false,
-    "strict-boolean-expressions": false,
-    "strict-type-predicates": false,
-    "switch-default": false,
-    "trailing-comma": false,
-    "typedef": false,
-    "variable-name": false
-  }
-}
